fix(azure services and tests): fixed tests and dependency injection so that AzureTest runs
This included injecting a nullLogger. Should probably replace this further down the line.
