--- src/Sepes.Infrastructure/Dto/AzureResourceDto.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Sepes.Infrastructure.Dto
-{
-    public class CloudResourceDto : UpdateableBaseDto
-    {
-        public string ResourceType { get; set; }
-    }
-}
