+++ src/Sepes.Infrastructure/Dto/CloudResourceDto.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Sepes.Infrastructure.Dto
+{
+    public class CloudResourceDto : UpdateableBaseDto
+    {
+        public string ResourceType { get; set; }
+    }
+}
