+++ src/Sepes.Infrastructure/Service/Azure/AzureVNetService.cs
@@ -1,5 +1,4 @@
-﻿
-using Microsoft.Azure.Management.Fluent;
+﻿using Microsoft.Azure.Management.Fluent;
 using Microsoft.Azure.Management.Network.Fluent;
 using Microsoft.Azure.Management.Network.Fluent.Models;
 using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
